chore(ui): introduce `BlueprintsNodeActions` - AB-103
